openssl: add EVP_PKEY methods for RSA DSA and EC keys
As associating EVP_PKEY methods with specifc keys require
the use of the engine API, the added methods are
activated in the next patch where a private engine for
internal use is added.

All changes are implemented in a new file named
pkcs11h-openssl-1.1.1.c which is meant to be selected at
build time if OPENSSL version is > 1.1.1 and engine
support is available. Autoconf changes are done
in a following commit.

Signed-off-by: Selva Nair <selva.nair@gmail.com>